MSF-89 - Scoping the flow.
